requirements.txt
python-telegram-bot==20.8
pytesseract
Pillow
python-dotenv